+++ Rebus.AzureServiceBus.Tests/AzureServiceBusPeekLockRenewalTest.cs
@@ -30,7 +30,8 @@ namespace Rebus.AzureServiceBus.Tests
 
         protected override void SetUp()
         {
-            _transport = new AzureServiceBusTransport(ConnectionString, QueueName, _consoleLoggerFactory, new TplAsyncTaskFactory(_consoleLoggerFactory));
+            _transport = new AzureServiceBusTransport(QueueName);
+            //_transport = new AzureServiceBusTransport(ConnectionString, QueueName, _consoleLoggerFactory, new TplAsyncTaskFactory(_consoleLoggerFactory));
 
             Using(_transport);