+++ Rebus.AzureServiceBus.Tests/NativeDeferTest.cs
@@ -30,7 +30,8 @@ namespace Rebus.AzureServiceBus.Tests
             var consoleLoggerFactory = new ConsoleLoggerFactory(false);
             var asyncTaskFactory = new TplAsyncTaskFactory(consoleLoggerFactory);
             var busLifetimeEvents = new BusLifetimeEvents();
-            new AzureServiceBusTransport(connectionString, QueueName, consoleLoggerFactory, asyncTaskFactory).PurgeInputQueue();
+            new AzureServiceBusTransport(QueueName).PurgeInputQueue();
+            //new AzureServiceBusTransport(connectionString, QueueName, consoleLoggerFactory, asyncTaskFactory).PurgeInputQueue();
 
             _activator = new BuiltinHandlerActivator();