+++ CHANGELOG.md
@@ -39,6 +39,10 @@
 * Remove ability to run on the "Basic" tier because it makes the world simpler, and you should at least be using "Standard" anyway
 * Add ability to set actual message ID on the `BrokeredMessage` by using Rebus' message ID as the value
 
+## 6.0.0-b01
+
+* Update to Microsoft's new driver and thus gain .NET Core support - finally!
+
 
 [lezzi]: https://github.com/lezzi
 [Meyce]: https://github.com/Meyce